feat(xv6-mbc): kernel.elf LINKS — RV32 conversion + 5 more iterations
ASCEND-LINUX Phase 1.1. Drove the build through 5 more iterations after
the first attempt; kernel/kernel.elf now LINKS GREEN. Translator runs;
remaining work is teaching it about CSR translation + .S register
restriction.

Iterations landed in this commit:

ITER 2 — RV64 → RV32 sd/ld in xv6 .S files
- adapters/swtch_mbc.S — context switch sd→sw, offsets halved.
- adapters/trampoline_mbc.S — 67 sd/ld → sw/lw, generated from upstream
  via Python script; offsets halved for RV32 trapframe layout.
- adapters/kernelvec_mbc.S — same rewrite for kernel trap vector;
  256-byte stack frame → 128.
- upstream/kernel/proc.h — patched: struct context u64×14 → u32×14 (56
  bytes); struct trapframe u64 → u32 with halved /* NN */ comments.
- Makefile.mbc OBJS: swtch.o → swtch_mbc.o; trampoline.o →
  trampoline_mbc.o; kernelvec.o → kernelvec_mbc.o.

ITER 3 — RV32 Sv32 virtual memory
- upstream/kernel/riscv.h — MAXVA: (1L<<38) → (1U<<31) for Sv32 32-bit
  VA. TRAMPOLINE/TRAPFRAME constants now fit in 32-bit immediates.

ITER 4 — Preprocessor on .S files
- Makefile.mbc: %.o: %.S now uses $(CC) (gcc) instead of $(AS), so
  trampoline.S/kernelvec.S get cpp expansion of memlayout.h's
  MAXVA-derived TRAPFRAME constant.

ITER 5 — Linker emulation + missing symbols
- Makefile.mbc LDFLAGS gain -melf32lriscv (was defaulting to elf64).
- adapters/kernel-mbc.ld provides PROVIDE(end = .) for xv6 kalloc
  to find heap start.
- upstream/kernel/main.c patched: plicinit/plicinithart/virtio_disk_init
  calls commented out (no PLIC; virtio handled by blk-ramdisk).

ITER 6 — libgcc soft-runtime stubs
- adapters/libgcc_stubs.c (NEW, 70 LOC): __udivdi3, __umoddi3 (64-bit
  divide for printf %ld), __sync_lock_test_and_set_4 / __sync_lock_release_4
  / __sync_synchronize (atomic builtins for spinlock — backed by MBC
  XCHG opcode 0x3D + FENCE 0x3F via inline-asm word emit).
- Makefile.mbc OBJS: + $A/libgcc_stubs.o.

ITER 7 — stack0 + uintptr_t
- adapters/start_mbc.c: declared `char stack0[4096*NCPU]` (entry.S
  expects it; was in upstream start.c we replaced).
- adapters/blk-ramdisk.c: removed (uintptr_t) cast (xv6 has no
  stdint.h; uint32 cast is sufficient for RV32).

RESULT: kernel/kernel.elf LINKS GREEN.
- 28 .o files compiled (entry, console, printf, kalloc, spinlock,
  string, main, vm, proc, trap, syscall, sysproc, bio, fs, log,
  sleeplock, file, pipe, exec, sysfile + 8 adapter .o files).
- Linker emits 32-bit ELF; section layout matches kernel-mbc.ld.
- objdump -d kernel/kernel.elf > kernel.asm — readable.

NEXT (Phase 1.1 final blockers):
- rv32i-to-mbc translator surface 56 x16+ register usages + 128 CSR
  opcodes that need translation. Two paths:
  (a) Teach translator (crates/monad-mbc/src/translator.rs) to:
      - Reject x16+ regs as compiler bug (we use -ffixed-x16..x31).
        Actually the .S adapter files use t0/t3-t6 = x5/x28-x31 since
        gas doesn't honor -ffixed-*. Rewrite .S files to use only
        registers in x0-x15 range (a0=x10, a1=x11, t0=x5 OK; t3-t6=
        x28-x31 NOT ok).
      - Emit memory-mapped LD/ST for CSRR* opcodes per ADR-067 Decision
        2: CSRR a, csr → LD a, 0xF000+csr*4. Roughly 50 LOC in
        translator.rs.
  (b) Strip CSR usage from xv6's .S files entirely; replace with
      C-level memory-mapped CSR access (slower but simpler).

Recommend (a). Estimated 1-2 days.

🎖️ Marshal handoff for next shift: invoke /unheaded-marshal with:
"ASCEND-LINUX P1.1 kernel.elf LINKS green (28 .o files). Translator
runs. Blocked on 56 x16+ regs + 128 CSR ops in compiled output. NEXT:
teach crates/monad-mbc/src/translator.rs to (1) emit memory-mapped
LD/ST for CSRR* per ADR-067 Decision 2 (~50 LOC), (2) flag x16+ regs
as a hard error so we catch missing -ffixed-* in .S files. Then
adapters/kernelvec_mbc.S + trampoline_mbc.S need t3-t6 → t0-t2 (or
s-regs) rewrites since those map to x28-x31. Owner Computermancer +
Developer."

+++ crates/xv6-mbc/adapters/blk-ramdisk.c
@@ -67,7 +67,7 @@ virtio_disk_rw(struct buf *b, int write)
     // xv6 block index → UPC sector indices. BSIZE=1024 = 2 × 512.
     uint32 sector0 = b->blockno * 2;
     uint32 sector1 = sector0 + 1;
-    uint32 buf_addr = (uint32)(uintptr_t)&b->data[0];
+    uint32 buf_addr = (uint32)&b->data[0];
     uint32 buf_addr_2nd = buf_addr + 512;
 
     if (write) {
